# Conflicts:
#	.travis.yml
#	Docs/AngelScriptAPI.h
#	Docs/LuaScriptAPI.dox
#	Docs/ScriptAPI.dox
#	Docs/Urho3D.dox
#	README.md
#	Rakefile
#	Source/CMakeLists.txt
#	Source/ThirdParty/LICENSES
#	Source/Tools/SpritePacker/SpritePacker.cpp
#	Source/Urho3D/.soversion
#	Source/Urho3D/AngelScript/Addons.cpp
#	Source/Urho3D/AngelScript/MathAPI.cpp
#	Source/Urho3D/AngelScript/ResourceAPI.cpp
#	Source/Urho3D/AngelScript/ScriptFile.cpp
#	Source/Urho3D/Core/Timer.cpp
#	Source/Urho3D/Graphics/Graphics.h
#	Source/Urho3D/Graphics/OpenGL/OGLGraphics.cpp
#	Source/Urho3D/Graphics/OpenGL/OGLGraphicsImpl.h
#	Source/Urho3D/Graphics/OpenGL/OGLTexture.cpp
#	Source/Urho3D/LuaScript/pkgs/Resource/JSONValue.pkg
#	Source/Urho3D/Resource/Decompress.cpp
#	Source/Urho3D/Resource/Decompress.h
#	Source/Urho3D/Resource/Image.cpp
#	Source/Urho3D/Resource/JSONValue.h
#	Source/Urho3D/Scene/Node.cpp
#	Source/Urho3D/Scene/Node.h
#	Source/Urho3D/UI/UI.cpp
#	build.gradle.kts

+++ Source/Urho3D/Core/Timer.cpp
@@ -99,7 +99,7 @@ static long long HiresTick()
     else
         return GetTickCount();
 #elif __EMSCRIPTEN__
-    return (unsigned)(emscripten_get_now()*1000.0);
+    return (long long)(emscripten_get_now()*1000.0);
 #else
     struct timeval time{};
     gettimeofday(&time, nullptr);
